+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.67
 ### is.codion.common.reactive
 - Observer.Builder added along with AbstractObserver.AbstractBuilder, related refactoring.
+- Observer.observer() added, Observable, default Observer method implementations moved to Observer, related refactoring.
 ### is.codion.common.db
 - Database.close() and AbstractDatabase.closeDatabase() added, for closing embedded databases.
 ### is.codion.common.model